requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
pyyaml>=6.0
requests>=2.31.0
python-telegram-bot>=21.0
